+++ nthSemesterGraduationPrediction/zeroRtest.py
@@ -1,0 +1,5 @@
+
+import numpy as np
+prediction_array = np.zeros(0)
+for x in range(1, 6):
+    print(x)
